feat(codegen): introduce module jump
